reel_auto/requirements.txt
requests
pandas
Pillow
moviepy
pydrive2
slack_sdk
croniter
streamlit
ffmpeg-python
